Fix columns on orders and orders product tables

+++ database/migrations/2020_06_20_223744_create_order_products_table.php
@@ -21,7 +21,7 @@ class CreateOrderProductsTable extends Migration
             $table->text("description")->nullable();
             $table->integer("price");
             $table->integer("quantity");
-            $table->integer("subtotal");
+            $table->float("subtotal");
         });
     }
